+++ wdsp/anr.hpp
@@ -48,19 +48,19 @@ public:
     int mask;
     int n_taps;
     int delay;
-    float two_mu;
-    float gamma;
-    float d [ANR_DLINE_SIZE];
-    float w [ANR_DLINE_SIZE];
+    double two_mu;
+    double gamma;
+    double d [ANR_DLINE_SIZE];
+    double w [ANR_DLINE_SIZE];
     int in_idx;
 
-    float lidx;
-    float lidx_min;
-    float lidx_max;
-    float ngamma;
-    float den_mult;
-    float lincr;
-    float ldecr;
+    double lidx;
+    double lidx_min;
+    double lidx_max;
+    double ngamma;
+    double den_mult;
+    double lincr;
+    double ldecr;
 
     static ANR* create_anr   (
         int run,
@@ -71,16 +71,15 @@ public:
         int dline_size,
         int n_taps,
         int delay,
-        float two_mu,
-        float gamma,
-
-        float lidx,
-        float lidx_min,
-        float lidx_max,
-        float ngamma,
-        float den_mult,
-        float lincr,
-        float ldecr
+        double two_mu,
+        double gamma,
+        double lidx,
+        double lidx_min,
+        double lidx_max,
+        double ngamma,
+        double den_mult,
+        double lincr,
+        double ldecr
     );
 
     static void destroy_anr (ANR *a);
@@ -91,11 +90,11 @@ public:
     static void setSize_anr (ANR *a, int size);
     // RXA Properties
     static void SetANRRun (RXA& rxa, int setit);
-    static void SetANRVals (RXA& rxa, int taps, int delay, float gain, float leakage);
+    static void SetANRVals (RXA& rxa, int taps, int delay, double gain, double leakage);
     static void SetANRTaps (RXA& rxa, int taps);
     static void SetANRDelay (RXA& rxa, int delay);
-    static void SetANRGain (RXA& rxa, float gain);
-    static void SetANRLeakage (RXA& rxa, float leakage);
+    static void SetANRGain (RXA& rxa, double gain);
+    static void SetANRLeakage (RXA& rxa, double leakage);
     static void SetANRPosition (RXA& rxa, int position);
 };
